requirements.txt
annotated-types==0.7.0
anyio==4.5.2
certifi==2025.8.3
click==8.1.8
colorama==0.4.6
deprecation==2.1.0
exceptiongroup==1.3.0
fastapi==0.117.1
gotrue==2.9.2
h11==0.16.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.9
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
numpy==1.24.4
packaging==25.0
pandas==2.0.3
postgrest==0.16.11
pydantic==2.10.6
pydantic_core==2.27.2
pykiwoom==0.1.6
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
pywin32==311
realtime==1.0.6
six==1.17.0
sniffio==1.3.1
starlette==0.44.0
storage3==0.7.7
StrEnum==0.4.15
supabase==2.6.0
supafunc==0.5.1
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.33.0
websockets==12.0
